simplify data model, enable download

+++ dfs/internal/datanode/resources.go
@@ -2,6 +2,6 @@ package datanode
 
 // System resources management
 
-func (s *DataNodeServer) hasCapacity(chunkSize int) bool {
+func (s *DataNodeServer) hasCapacity(chunkSize int64) bool {
 	return true
 }
